+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - updated arguments for datacenter, ipblock, location and user
 - issue #72 - fixed find volume image by name
+- error message for immutable node pool attributes
 
 ## 6.0.0-beta.5